fake_news_detector/requirement.txt
fastapi
uvicorn
transformers
sentence-transformers
newspaper3k
requests
torch
python-dotenv